Oh, here we go... I actually pushed the code now

+++ src/main/java/mc/Mitchellbrine/anvilModLoader/AML.java
@@ -1,10 +1,13 @@
 package mc.Mitchellbrine.anvilModLoader;
 
+import com.sun.javafx.fxml.expression.BinaryExpression;
 import mc.Mitchellbrine.anvilModLoader.database.*;
+import mc.Mitchellbrine.anvilModLoader.util.ANSIColor;
 import mc.Mitchellbrine.anvilModLoader.util.FileHelper;
 
 import net.minecraftforge.fml.relauncher.FMLInjectionData;
 import net.minecraftforge.fml.relauncher.IFMLLoadingPlugin;
+import org.apache.commons.codec.BinaryEncoder;
 
 //import cpw.fml.relauncher.FMLInjectionData;
 //import cpw.fml.relauncher.IFMLLoadingPlugin;
@@ -27,10 +30,17 @@ public class AML implements IFMLLoadingPlugin {
 
     public static ArrayList<ModDatabase> databases = new ArrayList<ModDatabase>();
 
+    private final String[] trustedSites = new String[]{"http://minecraft.curseforge.com","http://chickenbones.net","http://creeperrepo.net","http://planetminecraft.com","http://dl.dropboxusercontent.com/u/","http://copy.com", "https://github.com","https://bitbucket.org","http://ci.candicejoy.com:8080"};
+
+    private static boolean filterEnabled = true;
+
     public static AML instance;
 
     public AML() {
         instance = this;
+        new AMLDirectories();
+        AMLDirectories.mkDir();
+        instance.checkForConfig();
         instance.buildDatabase();
     }
 
@@ -59,13 +69,68 @@ public class AML implements IFMLLoadingPlugin {
         return null;
     }
 
+    private void checkForConfig() {
+        File mainDirectory = (File) AMLDirectories.data()[0];
+
+        File config = new File(mainDirectory,"config.cfg");
+
+        if (!config.exists()) {
+            try {
+                PrintWriter writer = new PrintWriter(config);
+                writer.println("#### The configuration file for AML ####");
+                writer.println("");
+                writer.println("# This config option allows you to disable AML's filter on trusted sites.");
+                writer.println("# Although you CANNOT hold myself, Mitchellbrine, accountable for your actions, as can be seen in the LGPLv2.1 license AML uses,");
+                writer.println("# No support or help of any kind will be given to those who disable the filter. I cannot be help liable and I do not want people ");
+                writer.println("# possibly destroying others' computers and blaming me. So, I have given fair warning before you decide to switch it off.");
+                writer.println("# -Mitchellbrine");
+                writer.println("");
+                writer.println("enableFilter=true");
+
+                writer.close();
+
+            } catch (IOException ex) {
+                ex.printStackTrace();
+            }
+        }
+
+
+
+            try {
+            BufferedReader reader = new BufferedReader(new InputStreamReader(new FileInputStream(config)));
+
+            String s;
+
+            while ((s = reader.readLine()) != null) {
+                if (s.startsWith("enableFilter=")) {
+                    filterEnabled = Boolean.parseBoolean(s.substring(s.indexOf("enableFilter=")));
+                    break;
+                }
+            }
+
+                reader.close();
+
+                if (!filterEnabled) {
+                    /*AMLProperties.logger.warn(ANSIColor.ANSI_RED + "YOU HAVE DISABLED THE FILTER OF TRUSTED SITES ON AML." + ANSIColor.ANSI_RESET);
+                    AMLProperties.logger.warn(ANSIColor.ANSI_RED + "Unfortunately, this means all support has been VOIDED " + ANSIColor.ANSI_RESET);
+                    AMLProperties.logger.warn(ANSIColor.ANSI_RED + "until you re-enable the filter. Sorry! -Mitchellbrine" + ANSIColor.ANSI_RESET); */
+                    //AMLProperties.logger.warn("0100011001010101010000100100000101010010");
+                    AMLProperties.logger.warn("AML Loading...");
+                } else {
+                    //AMLProperties.logger.warn("0100011001010101010000100100000101010011");
+                    AMLProperties.logger.info("AML Loading...");
+                }
+
+            } catch (IOException ex) {
+                ex.printStackTrace();
+            }
+
+    }
+
     /**
      * Build databases is the method we can call before any loading because it has nothing to do with Minecraft itself!
      */
     public void buildDatabase() {
-
-        new AMLDirectories();
-        AMLDirectories.mkDir();
 
         // TODO: Do all the work for building databases
 
@@ -87,6 +152,7 @@ public class AML implements IFMLLoadingPlugin {
 
                         boolean isCorrectVersion = false;
                         boolean isConfig = false;
+                        boolean isAbsolute = false;
 
                         if (s.startsWith("--") || !s.contains(":")) continue;
 
@@ -98,6 +164,8 @@ public class AML implements IFMLLoadingPlugin {
                         } else if (s.startsWith("+")) {
                             isConfig = true;
                             startingSubstring++;
+                        } else if (s.startsWith("$")) {
+                            isAbsolute = true;
                         }
 
                         int startingSubstring2 = s.indexOf(":") + 1;
@@ -106,7 +174,21 @@ public class AML implements IFMLLoadingPlugin {
                             startingSubstring2++;
                         }
 
-                        if (!isConfig) {
+                        if (!isConfig && !isAbsolute) {
+
+                            if (filterEnabled) {
+                                boolean isTrusted = false;
+
+                                for (String site : trustedSites) {
+                                    if (s.substring(startingSubstring2).startsWith(site)) {
+                                        isTrusted = true;
+                                    }
+                                }
+
+                                if (!isTrusted) continue;
+
+                            }
+
                             ModVersion version = new ModVersion(s.substring(startingSubstring, s.indexOf(":")), new URL(s.substring(startingSubstring2)));
 
                             database.getVersions().add(version);
@@ -115,16 +197,52 @@ public class AML implements IFMLLoadingPlugin {
                                 database.setCurrentVersion(version);
                             }
                             database.setDatabaseName(file.getName().substring(0, file.getName().lastIndexOf(".")));
+                        } else if (!isAbsolute){
+                            for (ModVersion version : database.getVersions()) {
+                                if (version.getVersion().equalsIgnoreCase(s.substring(startingSubstring, s.indexOf(":")))) {
+
+                                    if (filterEnabled) {
+                                        boolean isTrusted = false;
+
+                                        for (String site : trustedSites) {
+                                            if (s.substring(startingSubstring2).startsWith(site)) {
+                                                isTrusted = true;
+                                            }
+                                        }
+
+                                        if (!isTrusted) continue;
+
+                                    }
+
+                                    version.setConfigLocation(new URL(s.substring(startingSubstring2)));
+                                }
+                            }
                         } else {
                             for (ModVersion version : database.getVersions()) {
                                 if (version.getVersion().equalsIgnoreCase(s.substring(startingSubstring, s.indexOf(":")))) {
-                                    version.setConfigLocation(new URL(s.substring(startingSubstring2)));
+
+                                    if (filterEnabled) {
+                                        boolean isTrusted = false;
+
+                                        for (String site : trustedSites) {
+                                            if (s.substring(startingSubstring2).startsWith(site)) {
+                                                isTrusted = true;
+                                            }
+                                        }
+
+                                        if (!isTrusted) continue;
+
+                                    }
+
+                                    version.setAbsoluteLocation(new URL(s.substring(startingSubstring2)));
                                 }
                             }
                         }
 
 
                     }
+
+                    reader.close();
 
                     database.buildDirectories();
 
@@ -153,35 +271,66 @@ public class AML implements IFMLLoadingPlugin {
         for (File mod : modsDir.listFiles()) {
             if (mod == null) continue;
             if (mod.getName().contains("AnvilModLoader")) continue;
-            if (!mod.getName().endsWith(".zip") && !mod.getName().endsWith(".jar"))  { mod.delete(); continue; }
-
-            ZipFile zip = new ZipFile(mod);
-            Enumeration<? extends ZipEntry> entries = zip.entries();
-
-            boolean containsCoreMod = false;
-
-            while(entries.hasMoreElements()){
-                ZipEntry entry = entries.nextElement();
-                InputStream stream = zip.getInputStream(entry);
-
-                if (!entry.getName().endsWith(".MF")) continue;
-
-                BufferedReader reader = new BufferedReader(new InputStreamReader(stream));
-
-                String s;
-
-                while ((s = reader.readLine()) != null) {
-                    if (s.contains("FMLCorePlugin:")) {
-                        containsCoreMod = true;
-                    }
-                }
-
-                reader.close();
-
+            if (!mod.getName().endsWith(".zip") && !mod.getName().endsWith(".jar")) {
+                mod.delete();
+                continue;
             }
 
-            if (!containsCoreMod) {
-                mod.delete();
+            if (!mod.getName().contains("-")) continue;
+
+            int hyphenLoc = mod.getName().indexOf("-");
+
+            boolean hasRepo = false;
+
+            for (ModDatabase database : databases) {
+                if (database.getDatabaseName().equalsIgnoreCase(mod.getName().substring(0, hyphenLoc))) hasRepo = true;
+            }
+
+            if (hasRepo) {
+
+                ZipFile zip = new ZipFile(mod);
+                Enumeration<? extends ZipEntry> entries = zip.entries();
+
+                boolean containsCoreMod = false;
+
+                while (entries.hasMoreElements()) {
+                    ZipEntry entry = entries.nextElement();
+                    InputStream stream = zip.getInputStream(entry);
+
+                    if (!entry.getName().endsWith(".MF")) continue;
+
+                    BufferedReader reader = new BufferedReader(new InputStreamReader(stream));
+
+                    String s;
+
+                    while ((s = reader.readLine()) != null) {
+                        if (s.contains("FMLCorePlugin:")) {
+                            containsCoreMod = true;
+                        }
+                    }
+
+                    reader.close();
+
+                }
+
+                if (!containsCoreMod) {
+                    mod.delete();
+                } else {
+                    ModDatabase database = null;
+                    for (ModDatabase database1 : databases) {
+                        if (database1.getDatabaseName().equalsIgnoreCase(mod.getName().substring(0,hyphenLoc))) {
+                            database = database1;
+                        }
+                    }
+
+                    if (database != null) {
+                        if (database.getCurrentVersion() != null) {
+                            if (!mod.getName().substring(hyphenLoc + 1).equalsIgnoreCase(database.getCurrentVersion().toString())) {
+                                mod.delete();
+                            }
+                        }
+                    }
+                }
             }
         }
 
@@ -194,10 +343,10 @@ public class AML implements IFMLLoadingPlugin {
                     if (!newFileLoc.exists()) {
                         FileHelper.download(recommended.getFile().openStream(), 0, newFileLoc);
                     }
-                        FileHelper.copy(new FileInputStream(newFileLoc), new File((File) FMLInjectionData.data()[6], "/mods/" + database.getDatabaseName().toLowerCase() + "-" + recommended.getVersion() + ".jar"));
-                        if (new File((File)FMLInjectionData.data()[6],"/mods/" + database.getDatabaseName().toLowerCase() + "-" + recommended.getVersion() + ".jar").exists()) {
-                            AMLProperties.logger.info("Loaded version " + recommended.getVersion() + " from database " + database.getDatabaseName().toLowerCase());
-                        }
+                    FileHelper.copy(new FileInputStream(newFileLoc), new File((File) FMLInjectionData.data()[6], "/mods/" + database.getDatabaseName().toLowerCase() + "-" + recommended.getVersion() + ".jar"));
+                    if (new File((File)FMLInjectionData.data()[6],"/mods/" + database.getDatabaseName().toLowerCase() + "-" + recommended.getVersion() + ".jar").exists()) {
+                        AMLProperties.logger.info("Loaded version " + recommended.getVersion() + " from database " + database.getDatabaseName().toLowerCase());
+                    }
                     File configLocation = null;
                     if (recommended.getConfig() != null) {
                         configLocation = new File(recommended.getDirectory(),"/config/" + recommended.getConfig().getFile().lastIndexOf("/"));
@@ -211,6 +360,20 @@ public class AML implements IFMLLoadingPlugin {
                         FileHelper.copy(new FileInputStream(configLocation),new File((File) FMLInjectionData.data()[6], "/config/" + configLocation.getName()));
                         AMLProperties.logger.info("Loaded configs for " + recommended.getVersion() + " from database " + database.getDatabaseName().toLowerCase());
                     }
+                    File absoluteLocation = null;
+                    if (recommended.getAbsoluteFile() != null) {
+                        absoluteLocation = new File(recommended.getDirectory(), "" + recommended.getAbsoluteFile().getFile().substring(recommended.getAbsoluteFile().getFile().lastIndexOf("/")));
+
+                        if (!absoluteLocation.exists()) {
+                            absoluteLocation.mkdirs();
+                            FileHelper.download(recommended.getAbsoluteFile().openStream(), 0, absoluteLocation);
+                        }
+                    }
+
+                        if (absoluteLocation != null) {
+                            FileHelper.unzip(absoluteLocation,(File) FMLInjectionData.data()[6]);
+                            AMLProperties.logger.info("Loaded absolute for " + recommended.getVersion() + " from database" + database.getDatabaseName().toLowerCase());
+                        }
                 } catch (Exception ex) {
                     ex.printStackTrace();
                 }
@@ -237,6 +400,20 @@ public class AML implements IFMLLoadingPlugin {
                     if (configLocation != null) {
                         FileHelper.copy(new FileInputStream(configLocation),new File((File) FMLInjectionData.data()[6], "/config/" + configLocation.getName()));
                         AMLProperties.logger.info("Loaded configs for " + latest.getVersion() + " from database " + database.getDatabaseName().toLowerCase());
+                    }
+                    File absoluteLocation = null;
+                    if (latest.getAbsoluteFile() != null) {
+                        absoluteLocation = new File(latest.getDirectory(), "" + latest.getAbsoluteFile().getFile().substring(latest.getAbsoluteFile().getFile().lastIndexOf("/")));
+
+                        if (!absoluteLocation.exists()) {
+                            absoluteLocation.mkdirs();
+                            FileHelper.download(latest.getAbsoluteFile().openStream(), 0, absoluteLocation);
+                        }
+                    }
+
+                    if (absoluteLocation != null) {
+                        FileHelper.unzip(absoluteLocation,(File) FMLInjectionData.data()[6]);
+                        AMLProperties.logger.info("Loaded absolute for " + latest.getVersion() + " from database" + database.getDatabaseName().toLowerCase());
                     }
                 } catch (Exception ex) {
                     ex.printStackTrace();
